requirements.txt
flask==3.0.2
python-binance==1.0.19
pandas==2.2.1
pandas-ta==0.3.14b0
numpy==1.26.4
python-telegram-bot==20.8
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
TA-Lib==0.4.28
scipy==1.12.0
gunicorn==21.2.0
